Extract `FeedLoader` test helper in shared protocol extension to remove duplication
